LDE-714: moved tests for stats_cube
